test: move extends around to pass build
