+++ apps/main/main/src/test-rules/002-org.spec.ts
@@ -1,4 +1,4 @@
-import { setup, teardown } from './helpers';
+import { setup, teardown } from './helpers.spec';
 import { mockData, userGilles, userMax } from './mock';
 
 describe('Org rules', () => {
