+++ apps/main/main/src/test-rules/003-deals.spec.ts
@@ -1,4 +1,4 @@
-import { setup, teardown } from './helpers';
+import { setup, teardown } from './helpers.spec';
 import { mockData, userMax } from './mock';
 
 describe('Deals & Contracts rules', () => {
