+++ apps/main/main/src/test-rules/004-blockframes-admin.spec.ts
@@ -1,4 +1,4 @@
-import { setup, teardown } from './helpers';
+import { setup, teardown } from './helpers.spec';
 import { mockData, userGilles, userMarie, userMax, userVincentBlockframesAdmin } from './mock';
 
 describe('blockframes admins rules', () => {
