+++ apps/main/main/src/test-setup.ts
@@ -1,5 +1,3 @@
-import firebase from '@firebase/testing';
 import 'jest-preset-angular';
 import 'jest';
 import '../../../../../jestGlobalMocks';
-
